Endrett navn og relasjon på entitet
