+++ src/main/resources/db/migration/V150__opprett_overstyrt_utbetaling.sql
@@ -1,7 +1,7 @@
-CREATE TABLE OVERSTYRT_UTBETALING
+CREATE TABLE ENDRET_UTBETALING_ANDEL
 (
     ID                  BIGINT PRIMARY KEY,
-    FK_VEDTAK_ID        BIGINT REFERENCES VEDTAK (ID)       NOT NULL,
+    FK_BEHANDLING_ID    BIGINT REFERENCES BEHANDLING (ID)   NOT NULL,
     FK_PO_PERSON_ID     BIGINT REFERENCES PO_PERSON (ID)    NOT NULL,
     FOM                 TIMESTAMP(3)                        NOT NULL,
     TOM                 TIMESTAMP(3)                        NOT NULL,
@@ -15,6 +15,6 @@ CREATE TABLE OVERSTYRT_UTBETALING
     ENDRET_TID          TIMESTAMP(3)
 );
 
-CREATE SEQUENCE OVERSTYRT_UTBETALING_SEQ INCREMENT BY 50 START WITH 1000000 NO CYCLE;
-CREATE INDEX ON OVERSTYRT_UTBETALING (FK_VEDTAK_ID);
+CREATE SEQUENCE ENDRET_UTBETALING_ANDEL_SEQ INCREMENT BY 50 START WITH 1000000 NO CYCLE;
+CREATE INDEX ON ENDRET_UTBETALING_ANDEL (FK_BEHANDLING_ID);
